# daemon/api/swagger.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Minimal ping API server
paths:
  /board:
    post:
        description: Create board
        operationId: CreateBoard 
        requestBody:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewBoard"
        responses:
          '200':
            description: Ok
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Board'
  /board/{board}:
    parameters:
        - in: path
          name: board
          schema:
            type: string
          required: true
    post:
        description: Send message to board
        operationId: SendMessage
        requestBody:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        responses:
          '200':
            description: Ok
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Message'
  /subscription/{board}:
    parameters:
      - in: path
        name: board
        schema:
          type: string
        required: true
    post:
      description: Subscribe to board
      operationId: SubscribeBoard
      requestBody:
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/Subscribe'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
  /subscription/{board}/{user}:
      delete:
        description: Unsubscribe to board
        operationId: UnsubscribeBoard
        parameters:
        - in: path
          name: board
          schema:
            type: string
          required: true
        - in: path
          name: user
          schema:
            type: string
          required: true
        responses:
          '200':
            description: Ok
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Subscription'
components:
  schemas:
    Subscribe:
      type: object
      required:
        - notifiers
        - user
      properties:
        notifiers:
          type: array
          items: 
            type: string
        user:
          type: string
        password:
          type: string
    Subscription:
      type: object
      required:
        - notifiers
        - user
        - board
      properties:
        notifiers:
          type: array
          items: 
            type: string
        user:
          type: string
        password:
          type: string
        board:
          type: string
    Message:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
    NewBoard:
      type: object
      required:
        - board
      properties:
        board:
          type: string
        password:
          type: string
    Board:
      type: object
      required:
        - board
        - password
        - private
      properties:
        board:
          type: string
        password:
          type: string
        private:
          type: boolean
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string